Remove remaining wchar_t instantiation
